+++ src/main/java/com/example/todo/domain/repository/chat/ChatRoomRepository.java
@@ -1,7 +1,8 @@
 package com.example.todo.domain.repository.chat;
 
+import com.example.todo.domain.entity.TaskApiEntity;
 import com.example.todo.domain.entity.chat.ChatRoom;
-import com.example.todo.domain.entity.task.TaskApiEntity;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
